Recombine integration_tests in a single one.
